pyproject.toml
[project]
name = "dice-and-stats"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "customtkinter>=5.2.2",
    "dotenv>=0.9.9",
    "google-genai>=1.42.0",
    "ipykernel>=6.30.1",
    "python-dotenv>=1.1.1",
    "ttkthemes>=3.2.2",
]
